Changing occurence field to BigInteger

+++ words/models.py
@@ -6,7 +6,7 @@ from django.db import models
 class Word(models.Model):
 
     text = models.CharField(max_length=30)
-    occurence = models.IntegerField()
+    occurence = models.BigIntegerField()
 
     class Meta:
         app_label = 'words'
